Refactor friend cards to improve time to render
